Separating queue related functions to separate file
