+++ src/match/MatchMakingQueueItem.ts
@@ -1,5 +1,5 @@
 export type MatchMakingQueueItem = {
   username: string;
   timeJoinedAt: number;
-  playerRating: number;
+  playerRating?: number;
 };
